+++ cli/azd/pkg/infra/provisioning/manager.go
@@ -103,7 +103,7 @@ func (m *Manager) Deploy(ctx context.Context) (*DeployResult, error) {
 		m.console.StopSpinner(ctx, "Didn't find new changes.", input.StepSkipped)
 	}
 
-	if err := m.UpdateEnvironment(ctx, deployResult.Deployment.Outputs); err != nil {
+	if err := UpdateEnvironment(ctx, deployResult.Deployment.Outputs, m.env, m.envManager); err != nil {
 		return nil, fmt.Errorf("updating environment with deployment outputs: %w", err)
 	}
 
@@ -314,9 +314,11 @@ func (m *Manager) Destroy(ctx context.Context, options DestroyOptions) (*Destroy
 	return destroyResult, nil
 }
 
-func (m *Manager) UpdateEnvironment(
+func UpdateEnvironment(
 	ctx context.Context,
 	outputs map[string]OutputParameter,
+	env *environment.Environment,
+	envManager environment.Manager,
 ) error {
 	if len(outputs) > 0 {
 		for key, param := range outputs {
@@ -326,13 +328,13 @@ func (m *Manager) UpdateEnvironment(
 				if err != nil {
 					return fmt.Errorf("invalid value for output parameter '%s' (%s): %w", key, string(param.Type), err)
 				}
-				m.env.DotenvSet(key, string(bytes))
+				env.DotenvSet(key, string(bytes))
 			} else {
-				m.env.DotenvSet(key, fmt.Sprintf("%v", param.Value))
+				env.DotenvSet(key, fmt.Sprintf("%v", param.Value))
 			}
 		}
 
-		if err := m.envManager.Save(ctx, m.env); err != nil {
+		if err := envManager.Save(ctx, env); err != nil {
 			return fmt.Errorf("writing environment: %w", err)
 		}
 	}